+++ example_code/collision.h
@@ -1,3 +1,5 @@
+#pragma once
+
 void entity_DefaultCollide(Entity* e, Entity* with);
 void entity_IcebergCollide(Entity* e, Entity* with);
 void entity_SharkCollide(Entity* e, Entity* with);
