+++ example_code/draw.h
@@ -1,2 +1,4 @@
+#pragma once
+
 void DrawGame();
 void DrawMenu();
