+++ example_code/input.h
@@ -1,3 +1,4 @@
+#pragma once
 
 void MouseMotion(int x, int y);
 void OnMouseClick(int button, int state, int x, int y);
