+++ example_code/linkedlist.cpp
@@ -1,4 +1,6 @@
-#include "main.h"
+
+#include <stdlib.h>
+#include "linkedlist.h"
 
 entlist* list_Create() {
   entlist* a = (entlist*)calloc(1, sizeof(entlist));
@@ -40,22 +42,21 @@ bool list_Remove(Entity* e, entlist* lst) {
   }
 }
 
-bool list_Add(Entity* e, entlist* lst) {
-  if (!lst) return false;
+Entity* list_Add(Entity e, entlist* lst) {
+  if (!lst) return (Entity*)(0);
 
-  listItem* lI = (listItem*)calloc(1, sizeof(listItem));
-  lI->data = *e;
-  lI->next=(listItem*)(0);
+  listItem* lstItem = (listItem*)calloc(1, sizeof(listItem));
+  lstItem->data = e;
+  lstItem->next = (listItem*)(0);
 
-  if (lst->last!=(listItem*)0){
-	lst->last->next=lI;
-	lst->last=lI;
+  if (lst->last != (listItem*)0){
+    lst->last->next = lstItem;
+    lst->last = lstItem;
+  } else {
+    lst->first = lstItem;
+    lst->last = lstItem;
   }
-  else{
-	lst->first=lI;
-	lst->last=lI;
-  }
-
+  return &lstItem->data;
 }
