+++ example_code/linkedlist.h
@@ -1,3 +1,5 @@
+#pragma once
+#include "entity.h"
 
 typedef struct listItem {
   listItem* next;
@@ -12,4 +14,4 @@ typedef struct entlist {
 entlist* list_Create();
 bool list_Destroy(entlist* lst);
 bool list_Remove(Entity* e, entlist* lst);
-bool list_Add(Entity* e, entlist* lst);
+Entity* list_Add(Entity* e, entlist* lst);
