+++ example_code/mouse.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "coords.h"
+
+typedef struct Mouse {
+  Coords coords;
+  bool btn1;
+  bool btn2;
+} Mouse;
