+++ example_code/movement.cpp
@@ -1,4 +1,7 @@
-#include "main.h"
+#include "entity.h"
+#include "gamestate.h"
+#include <math.h>
+#include "movement.h"
 
 void entity_DefaultMove(Entity* e) {
   e->coords.x += e->movement.x;
