+++ example_code/update.h
@@ -1,1 +1,3 @@
+#pragma once
+
 void Update(int t);
